todo_project/requirements.txt
django==3.2.25
djangorestframework==3.12.4
psycopg2-binary==2.9
celery==5.2
redis==4.0
